# packages/sdk/openapi/v1.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/players/{idOrNick}": {
      "get": {
        "operationId": "getPlayer",
        "parameters": [
          {
            "name": "idOrNick",
            "required": true,
            "in": "path",
            "schema": {
              "minLength": 2,
              "maxLength": 24,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1PlayerProfileDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "API_KEY_INVALID or API_KEY_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "PLAYER_NOT_FOUND, CLAN_NOT_FOUND, TANK_NOT_FOUND or NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED or PLAN_LIMIT_REACHED; see the Retry-After header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "summary": "Player summary with WN8, EFF, Броня-Индекс and every recent period",
        "tags": ["v1-players"]
      }
    },
...
    "/v1/players/{id}/tanks": {
      "get": {
        "operationId": "listPlayerTanks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": true,
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "tiers",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 1,
                "maximum": 11
              }
            }
          },
          {
            "name": "types",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["lightTank", "mediumTank", "heavyTank", "AT-SPG", "SPG"]
              }
            }
          },
          {
            "name": "nations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 32
              }
            }
          },
          {
            "name": "premium",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "collectible",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["battles", "winRate", "avgDamage", "wn8", "marksOnGun", "moePercent", "lastBattleAt", "tier"]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Page size up to 1000, or \"all\" for every tank in one page",
            "schema": {
              "default": 25,
              "anyOf": [
                {
                  "type": "string",
                  "enum": ["all"]
                },
                {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 1000
                }
              ]
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0,
              "type": "integer"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["1000", "24h", "7d", "30d", "60d"]
            }
          },
          {
            "name": "minBattles",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1PlayerTanksPageDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "API_KEY_INVALID or API_KEY_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "PLAYER_NOT_FOUND, CLAN_NOT_FOUND, TANK_NOT_FOUND or NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED or PLAN_LIMIT_REACHED; see the Retry-After header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "summary": "Per-tank statistics",
        "tags": ["v1-players"]
      }
    },
...
    "/v1/tanks": {
      "get": {
        "operationId": "listTankStats",
        "parameters": [
          {
            "name": "tiers",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "minimum": 1,
                "maximum": 11
              }
            }
          },
          {
            "name": "types",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["lightTank", "mediumTank", "heavyTank", "AT-SPG", "SPG"]
              }
            }
          },
          {
            "name": "nations",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "minLength": 1,
                "maxLength": 32
              }
            }
          },
          {
            "name": "premium",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "collectible",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "statuses",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["researchable", "premium", "collector", "reward", "removed"]
              }
            }
          },
          {
            "name": "roles",
            "required": false,
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "HT_assault",
                  "HT_break",
                  "HT_universal",
                  "HT_support",
                  "MT_universal",
                  "MT_sniper",
                  "MT_assault",
                  "MT_support",
                  "LT_universal",
                  "LT_wheeled",
                  "ATSPG_sniper",
                  "ATSPG_assault",
                  "ATSPG_support",
                  "ATSPG_universal",
                  "SPG",
                  "SPG_assault",
                  "SPG_flame"
                ]
              }
            }
          },
          {
            "name": "difficulties",
            "required": false,
            "in": "query",
            "description": "Only tanks whose learning curve puts them in one of these difficulties",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": ["easy", "moderate", "hard", "hardcore"]
              }
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["battles", "players", "winRate", "winRateDiff", "avgDamage", "avgFrags", "avgSpotted", "survivalRate", "accuracy", "tier"]
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": ["asc", "desc"]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 25,
              "type": "integer"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0,
              "type": "integer"
            }
          },
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "7d",
              "type": "string",
              "enum": ["1d", "7d", "14d", "30d", "60d"]
            }
          },
          {
            "name": "cohort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "all",
              "type": "string",
              "enum": ["all", "beginner", "average", "good", "elite"]
            }
          },
          {
            "name": "mode",
            "required": false,
            "in": "query",
            "schema": {
              "default": "random",
              "type": "string",
              "enum": ["all", "random", "ranked", "epic", "clan", "stronghold"]
            }
          },
          {
            "name": "minBattles",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1TankStatsPageDto_Output"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_FAILED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "API_KEY_INVALID or API_KEY_REVOKED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "PLAYER_NOT_FOUND, CLAN_NOT_FOUND, TANK_NOT_FOUND or NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED or PLAN_LIMIT_REACHED; see the Retry-After header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V1ApiErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "summary": "Server-wide statistics per tank: win rate, WR diff, damage, popularity",
        "tags": ["v1-tanks"]
      }
    },
...
  },
  "info": {
    "title": "Three Marks Public API",
    "description": "Derived «Мир танков» (Lesta RU) data: WN8 and recent periods, player history, server tank statistics, tier list, marks of excellence thresholds and their history, clans, leaderboards, sessions. Pass your key in the X-API-Key header; create one at /me/developer. Data source: Леста Игры.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "V1PlayerProfileDto_Output": {
        "type": "object",
        "properties": {
          "summary": {
            "type": "object",
            "properties": {
              "accountId": {
                "type": "integer",
                "exclusiveMinimum": true,
                "maximum": 9007199254740991,
                "minimum": 0
              },
              "nickname": {
                "type": "string"
              },
              "clan": {
                "type": "object",
                "properties": {
                  "clanId": {
                    "type": "integer",
                    "exclusiveMinimum": true,
                    "maximum": 9007199254740991,
                    "minimum": 0
                  },
                  "tag": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "color": {
                    "type": ["string", "null"]
                  },
                  "role": {
                    "type": "string"
                  },
                  "emblem": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "joinedAt": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                    "nullable": true
                  }
                },
                "required": ["clanId", "tag", "name", "color", "role", "emblem", "joinedAt"],
                "additionalProperties": false,
                "nullable": true
              },
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "nullable": true
              },
              "lastBattleAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                "nullable": true
              },
              "updatedAt": {
                "type": "string",
                "format": "date-time",
                "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$"
              },
              "isTracked": {
                "type": "boolean"
              },
              "overall": {
                "type": "object",
                "properties": {
                  "battles": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "winRate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "description": "Percent, 0–100",
                    "nullable": true
                  },
                  "avgDamage": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true
                  },
                  "avgFrags": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true
                  },
                  "avgSpotted": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true
                  },
                  "avgXp": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true
                  },
                  "avgBlocked": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true
                  },
                  "avgAssisted": {
                    "type": "number",
                    "minimum": 0,
                    "nullable": true
                  },
                  "survivalRate": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "description": "Percent, 0–100",
                    "nullable": true
                  },
                  "accuracy": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 100,
                    "description": "Percent, 0–100",
                    "nullable": true
                  },
                  "avgTier": {
                    "type": "number",
                    "minimum": 1,
                    "maximum": 11,
                    "nullable": true
                  },
                  "wn8": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": ["number", "null"]
                      },
                      "tier": {
                        "type": "string",
                        "enum": ["very_bad", "bad", "below_avg", "avg", "good", "very_good", "great", "unicum", "super_unicum", null],
                        "nullable": true
                      }
                    },
                    "required": ["value", "tier"],
                    "additionalProperties": false
                  },
                  "eff": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": ["number", "null"]
                      },
                      "tier": {
                        "type": "string",
                        "enum": ["very_bad", "bad", "below_avg", "avg", "good", "very_good", "great", "unicum", "super_unicum", null],
                        "nullable": true
                      }
                    },
                    "required": ["value", "tier"],
                    "additionalProperties": false
                  },
                  "broneIndex": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": ["number", "null"]
                      },
                      "tier": {
                        "type": "string",
                        "enum": ["very_bad", "bad", "below_avg", "avg", "good", "very_good", "great", "unicum", "super_unicum", null],
                        "nullable": true
                      }
                    },
                    "required": ["value", "tier"],
                    "additionalProperties": false
                  }
                },
                "required": [
                  "battles",
                  "winRate",
                  "avgDamage",
                  "avgFrags",
                  "avgSpotted",
                  "avgXp",
                  "avgBlocked",
                  "avgAssisted",
                  "survivalRate",
                  "accuracy",
                  "avgTier",
                  "wn8",
                  "eff",
                  "broneIndex"
                ],
                "additionalProperties": false
              },
              "marks": {
                "type": "object",
                "properties": {
                  "moe3": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "moe2": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "moe1": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "mastery": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "tanksOwned": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": ["moe3", "moe2", "moe1", "mastery", "tanksOwned"],
                "additionalProperties": false
              }
            },
            "required": ["accountId", "nickname", "clan", "createdAt", "lastBattleAt", "updatedAt", "isTracked", "overall", "marks"],
            "additionalProperties": false
          },
          "recent": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "period": {
                  "type": "string",
                  "enum": ["1000", "24h", "7d", "30d", "60d"]
                },
                "from": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                  "nullable": true
                },
                "to": {
                  "type": "string",
                  "format": "date-time",
                  "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
                  "nullable": true
                },
                "stats": {
                  "type": "object",
                  "properties": {
                    "battles": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "winRate": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100,
                      "description": "Percent, 0–100",
                      "nullable": true
                    },
                    "avgDamage": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true
                    },
                    "avgFrags": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true
                    },
                    "avgSpotted": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true
                    },
                    "avgXp": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true
                    },
                    "avgBlocked": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true
                    },
                    "avgAssisted": {
                      "type": "number",
                      "minimum": 0,
                      "nullable": true
                    },
                    "survivalRate": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100,
                      "description": "Percent, 0–100",
                      "nullable": true
                    },
                    "accuracy": {
                      "type": "number",
                      "minimum": 0,
                      "maximum": 100,
                      "description": "Percent, 0–100",
                      "nullable": true
                    },
                    "avgTier": {
                      "type": "number",
                      "minimum": 1,
                      "maximum": 11,
                      "nullable": true
                    },
                    "wn8": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": ["number", "null"]
                        },
                        "tier": {
                          "type": "string",
                          "enum": ["very_bad", "bad", "below_avg", "avg", "good", "very_good", "great", "unicum", "super_unicum", null],
                          "nullable": true
                        }
                      },
                      "required": ["value", "tier"],
                      "additionalProperties": false
                    },
                    "eff": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": ["number", "null"]
                        },
                        "tier": {
                          "type": "string",
                          "enum": ["very_bad", "bad", "below_avg", "avg", "good", "very_good", "great", "unicum", "super_unicum", null],
                          "nullable": true
                        }
                      },
                      "required": ["value", "tier"],
                      "additionalProperties": false
                    },
                    "broneIndex": {
                      "type": "object",
                      "properties": {
                        "value": {
                          "type": ["number", "null"]
                        },
                        "tier": {
                          "type": "string",
                          "enum": ["very_bad", "bad", "below_avg", "avg", "good", "very_good", "great", "unicum", "super_unicum", null],
                          "nullable": true
                        }
                      },
                      "required": ["value", "tier"],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "battles",
                    "winRate",
                    "avgDamage",
                    "avgFrags",
                    "avgSpotted",
                    "avgXp",
                    "avgBlocked",
                    "avgAssisted",
                    "survivalRate",
                    "accuracy",
                    "avgTier",
                    "wn8",
                    "eff",
                    "broneIndex"
                  ],
                  "additionalProperties": false,
                  "nullable": true
                }
              },
              "required": ["period", "from", "to", "stats"],
              "additionalProperties": false
            }
          }
        },
        "required": ["summary", "recent"],
        "additionalProperties": false
      },
...
    }
  }
}